Add definition activity and refactor activity main
