+++ app/src/main/java/edu/cs371m/visionary/MainActivity.kt
@@ -1,15 +1,13 @@
 package edu.cs371m.visionary
 
+import android.content.Intent
 import android.os.Bundle
-import android.view.View
 import android.view.inputmethod.InputMethodManager
 import android.widget.Toast
 import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
-import androidx.recyclerview.widget.LinearLayoutManager
 import edu.cs371m.visionary.databinding.ActivityMainBinding
 import edu.cs371m.visionary.databinding.ContentMainBinding
-import edu.cs371m.visionary.ui.DefinitionAdapter
 import edu.cs371m.visionary.ui.MainViewModel
 
 class MainActivity : AppCompatActivity() {
@@ -24,41 +22,19 @@ class MainActivity : AppCompatActivity() {
         imm.hideSoftInputFromWindow(window.decorView.rootView.windowToken, 0)
     }
 
-    private fun initRecyclerView(binding: ContentMainBinding) : DefinitionAdapter {
-        val definitionAdapter = DefinitionAdapter(viewModel)
-        val recyclerView = binding.recyclerView
-
-        recyclerView.adapter = definitionAdapter
-        recyclerView.layoutManager = LinearLayoutManager(this)
-
-        return definitionAdapter
-    }
-
     private fun initSearchDefinition() {
         binding.searchButton.setOnClickListener {
             hideKeyboard()
 
-            binding.definition.text = null
             val word = binding.plainTextInput.text.toString()
 
             if (word.isBlank()) {
                 val message = "Please enter a word into the text box"
                 Toast.makeText(this, message, Toast.LENGTH_SHORT).show()
             } else {
-                viewModel.setWord(word)
-                viewModel.netDefinitions()
-            }
-        }
-    }
-
-    private fun initClearDefinition() {
-        binding.clear.setOnClickListener {
-            if (binding.clear.visibility == View.VISIBLE) {
-                binding.definition.text = null
-                binding.plainTextInput.text = null
-                binding.clear.visibility = View.INVISIBLE
-                binding.recyclerView.visibility = View.INVISIBLE
-                binding.click.visibility = View.INVISIBLE
+                val intent = Intent(it.context, DefinitionActivity::class.java)
+                intent.putExtra("word", word)
+                it.context.startActivity(intent)
             }
         }
     }
@@ -67,28 +43,8 @@ class MainActivity : AppCompatActivity() {
         super.onCreate(savedInstanceState)
         val activityMainBinding = ActivityMainBinding.inflate(layoutInflater)
         setContentView(activityMainBinding.root)
-
         binding = activityMainBinding.contentMain
 
-        val definitionAdapter = initRecyclerView(binding)
-
         initSearchDefinition()
-        initClearDefinition()
-
-        viewModel.observeMeaningDefinitions().observe(this) {
-            if(it.isNullOrEmpty()) {
-                // if the word has no definition
-                binding.definition.text = "There are no definitions for this word."
-                binding.click.visibility = View.INVISIBLE
-                binding.recyclerView.visibility = View.INVISIBLE
-            } else {
-                definitionAdapter.submitList(it)
-                binding.click.visibility = View.VISIBLE
-                binding.recyclerView.visibility = View.VISIBLE
-            }
-
-            // Change clear and search buttons visibility
-            binding.clear.visibility = View.VISIBLE
-        }
     }
 }
